Tower Placement and Game Manager
Various changes to tower placement and Game Manager
