+++ Assets/Materials/TransparentMat.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2195aff5be6d73145a0447cd50f7f3b2
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
