+++ Assets/Prefabs/TowerPlacePre.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ebb68882b3f8e6d438a6fe66fd4cb0ba
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
